split feed into smaller components
