+++ client/src/feed/Feed.jsx
@@ -4,8 +4,7 @@ import gql from 'graphql-tag';
 import AuthRoute from '../common/utils/AuthRoute';
 
 import CreatePost from '../common/CreatePost';
-
-import PrayerCard from '../common/PrayerCard/PrayerCard';
+import FeedCards from './FeedCards';
 
 export const GET_POSTS = gql`
   query getPostFeed($limit: Int, $offset: Int) {
@@ -18,30 +17,6 @@ export const GET_POSTS = gql`
     }
   }
 `;
-
-const buildPostCards = (posts, refetch) => {
-  const userId = localStorage.getItem('userId');
-
-  return posts.map(post => {
-    let isOwn = false;
-    if (post.owner === userId) {
-      isOwn = true;
-    }
-    return (
-      <PrayerCard
-        key={post.id}
-        id={post.id}
-        ownerId={post.owner}
-        isOwnCard={isOwn}
-        isPartnered={post.isPartnered}
-        partneredAmount={post.partnerCount}
-        postModifiedHandler={refetch}
-      >
-        {post.content}
-      </PrayerCard>
-    );
-  });
-};
 
 export default class Feed extends Component {
   constructor(props) {
@@ -75,57 +50,52 @@ export default class Feed extends Component {
     });
   }
 
-  cardsList() {
-    return (
-      <Query
-        query={GET_POSTS}
-        variables={{
-          limit: 10,
-          offset: 0,
-        }}
-        fetchPolicy="cache-and-network"
-      >
-        {({ loading, error, data, refetch, fetchMore }) => {
-          if (data && Object.keys(data).length !== 0) {
-            const posts = data.getPostFeed;
-
-            if (posts.length === 0) {
-              return (
-                <div>
-                  <h3>No posts in your feed!</h3>
-                  <p>Try adding some friends!</p>
-                </div>
-              );
-            }
-
-            const postCards = buildPostCards(posts, refetch);
-
-            return (
-              <React.Fragment>
-                {postCards}
-                {this.state.canFetchMore && (
-                  <button
-                    className="colored full-width"
-                    onClick={() => this.loadMore(fetchMore, posts)}
-                  >
-                    Load More
-                  </button>
-                )}
-              </React.Fragment>
-            );
-          }
-
-          if (loading) return <PrayerCard>Loading...</PrayerCard>;
-          return <p>Uh oh. An error has occured :(</p>;
-        }}
-      </Query>
-    );
-  }
-
   render() {
     return (
       <div>
-        <div className="MainView">{this.cardsList()}</div>
+        <div className="MainView">
+          <Query
+            query={GET_POSTS}
+            variables={{
+              limit: 10,
+              offset: 0,
+            }}
+            fetchPolicy="cache-and-network"
+          >
+            {({ loading, error, data, refetch, fetchMore }) => {
+              if (data && Object.keys(data).length !== 0) {
+                const posts = data.getPostFeed;
+
+                if (posts.length === 0) {
+                  return (
+                    <div>
+                      <h3>No posts in your feed!</h3>
+                      <p>Try adding some friends!</p>
+                    </div>
+                  );
+                }
+
+                return (
+                  <React.Fragment>
+                    <FeedCards>{posts}</FeedCards>
+                    {this.state.canFetchMore && (
+                      <button
+                        className="colored full-width"
+                        onClick={() => this.loadMore(fetchMore, posts)}
+                      >
+                        Load More
+                      </button>
+                    )}
+                  </React.Fragment>
+                );
+              }
+
+              if (loading) return <PrayerCard>Loading...</PrayerCard>;
+              return <p>Uh oh. An error has occured :(</p>;
+            }}
+          </Query>
+        </div>
+
         <AuthRoute exact path="/feed/create-post" component={CreatePost} />
       </div>
     );
